Fix call statements to lookup functions correctly.
